Move LiricalAnalysisException to `core.analysis` package.
