+++ lirical-core/src/main/java/org/monarchinitiative/lirical/core/analysis/LiricalAnalysisRunner.java
@@ -1,7 +1,5 @@
 package org.monarchinitiative.lirical.core.analysis;
 
-import org.monarchinitiative.lirical.core.exception.LiricalAnalysisException;
-
 import java.io.Closeable;
 
 /**
